Added addTask method
This method takes input from the user, checks if such task exists in HashSet. If not, creates new task. If category doesn't exist, creates a category with ArrayList of Task objects. Then adds task to ArrayList mapped to the category. Then it adds it to PriorityQueue for sorting. Then adds a notification.

I also changed the way PriorityQueue works by adding a custom comparator.